tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/004.html

<!DOCTYPE html>
<html><head>
  <title> scheduler: external+inline in markup </title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="testlib/testlib.js"></script>
</head>
<body>

  <div id="log">FAILED (This TC requires JavaScript enabled)</div>

  <script src="scripts/include-1.js"></script>
  <script>log('inline script #2')</script>

  <script type="text/javascript">
  var t = async_test()

  function test() {
    assert_array_equals(eventOrder, ['external script #1', 'inline script #2' ]);
    t.done();
  }
  onload = t.step_func(test)
  </script>

</body></html>
